ytb_dl/requirements.txt
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
mutagen==1.47.0
pycryptodomex==3.21.0
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
websockets==13.1
yt-dlp==2025.12.8
